Load entire workspace into store
Instead of just loading the direct dependencies from the
current Project.toml, also look at docs, test, extensions and
workspace projects.

Unfortunately, for extensions we need to resolve the
environment first. Test also needs to be resolved unless
it's included in the workspace configuration in which case
the project's manifest already contains the resolved dependencies.

Additionally, when running the server add the current project to the
list of packages to load, otherwise the loaded Symbol outside of
the package, e.g., in test/docs/etc., corresponds to the __stored__ module.
Ideally, LanguageServer.jl should detect with some limits when to re-index
the current project, but for now just using the "re-index" action works OK.